requirements.txt
Django==3.2
django-environ==0.4.5
# psycopg2>=2.8.6,<2.9
uWSGI>=2.0.19.1,<2.1
Pillow >= 9.0.0,<10.0.0
django-admin-interface == 0.19.1
djangorestframework == 3.13.1
django-smart-selects == 1.5.9
# django-jazzmin == 2.5.0
django-import-export == 2.8.0
django-ckeditor == 6.5.1
django-cors-headers
# django-dotenv